Refactor: extract shared alignment primitives into alexandria_alignment.py
Move ~700 lines of alignment + source-loading + lexicon code out of
alexandria_compare.py into a new shared module. The preparer
(alexandria_preparer_rocm_compatible.py) will import from this module to
implement source-guided chunking in the next commit. Zero behavior change
for the compare workflow.

What moved to alexandria_alignment.py:
  * Source loaders (load_epub, load_source) + optional EPUB import.
  * Source-load cleanups (_OCR_DIGIT_GLITCH, _DIACRITIC_REJOIN,
    _DIACRITIC_REJOIN_TAIL) plus new clean_source_text() convenience
    wrapper that applies all of them in one call.
  * Text normalisation (normalize, to_words, _expand_honorifics,
    strip_annotations, _SMART_QUOTES, _HONORIFICS).
  * Proper-noun lexicon (_PROPER_NOUNS, _build_proper_nouns,
    _PROPER_NOUN_STOPWORDS) + the three-tier threshold constants
    (_FUZZY_KEEP_THRESHOLD, _LONG, _PROPER_NOUN) and _step_threshold().
  * Number-word equivalence (_parse_number, _num_eq, _num_eq_step_*,
    _token_looks_numeric, _NUM_* dicts).
  * Fuzzy alignment helpers (_char_sim with lru_cache, _ratio, _span_size,
    _one_to_N_*, _two_by_two_*, _lookahead_anchor_*, trim_span_to_alignment).
  * Window search (find_best_match, find_anchor_position, auto_anchor,
    realign, estimate_alignment_quality, find_text_in_source).

What stayed in alexandria_compare.py:
  * load_jsonl (compare-specific).
  * parse_annotated_tokens + merge_annotations_with_source (the merge
    preview that re-applies LLM prosody markers onto source spelling —
    only meaningful in the review context).
  * color_diff, fmt_time (UI).
  * Checkpoint, review log, targeted reset, write_output, run, main.

main() now sets the per-book lexicon via _alignment._PROPER_NOUNS = ...
rather than a local `global` so the helpers in the shared module see the
mutation. Empty lexicon = pre-feature behaviour when used as a library.

Verified end-to-end: real-book lexicon build (29 entries), proper-noun trim
relaxation, compare merge (multi-word emphasis), parse_annotated_tokens
(fused-punct fix), _char_sim cache, clean_source_text (both diacritic
shapes), find_best_match across boundaries, CLI --help.
